+++ inference-engine/src/transformations/include/transformations/depth_to_space_fusion.hpp
